refactor: useChat custom hook
